# api/openapi/schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Social Theory Library API",
        "version": "3.0.7",
        "description": "Canonical catalog, editorial processes and permission-scoped reading APIs."
    },
    "paths": {
...
        "/api/catalog/admin/recommendation-issues/": {
            "get": {
                "operationId": "catalog_admin_recommendation_issues_list",
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "libraryCookie": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminRecommendationIssueCollection"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "catalog_admin_recommendation_issues_create",
                "tags": [
                    "catalog"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecommendationIssueRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RecommendationIssueRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RecommendationIssueRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "libraryCookie": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecommendationIssue"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/catalog/recommendation-issues/": {
            "get": {
                "operationId": "catalog_recommendation_issues_list",
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "libraryCookie": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecommendationIssueCollection"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminIssueSummary": {
                "type": "object",
                "properties": {
                    "total": {
                        "type": "integer"
                    },
                    "published": {
                        "type": "integer"
                    },
                    "drafts": {
                        "type": "integer"
                    },
                    "scheduled": {
                        "type": "integer"
                    }
                },
                "required": [
                    "drafts",
                    "published",
                    "scheduled",
                    "total"
                ]
            },
            "AdminRecommendationIssueCollection": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true
                    },
                    "current": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RecommendationIssue"
                            }
                        ],
                        "nullable": true
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecommendationIssue"
                        }
                    },
                    "summary": {
                        "$ref": "#/components/schemas/AdminIssueSummary"
                    },
                    "upcoming": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecommendationIssue"
                        }
                    }
                },
                "required": [
                    "count",
                    "current",
                    "next",
                    "previous",
                    "results",
                    "summary",
                    "upcoming"
                ]
            },
...
            "IssueBodyBlock": {
                "type": "object",
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/TypeEnum"
                    },
                    "text": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "required": [
                    "text",
                    "type"
                ]
            },
            "IssueBodyBlockRequest": {
                "type": "object",
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/TypeEnum"
                    },
                    "text": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "required": [
                    "text",
                    "type"
                ]
            },
            "IssueItem": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "kind": {
                        "$ref": "#/components/schemas/IssueItemKindEnum"
                    },
                    "work_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "edition_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "title": {
                        "type": "string"
                    },
                    "authors": {
                        "type": "string"
                    },
                    "version_note": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "doi": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "position": {
                        "type": "integer"
                    },
                    "document_type": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "readOnly": true
                    },
                    "work_url": {
                        "type": "string",
                        "readOnly": true
                    },
                    "reader_url": {
                        "type": "string",
                        "readOnly": true
                    },
                    "cover_url": {
                        "type": "string",
                        "readOnly": true
                    },
                    "file_status": {
                        "type": "string",
                        "readOnly": true
                    },
                    "cataloging_session_id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "nullable": true
                    },
                    "available_work_id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "nullable": true
                    },
                    "available_edition_id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "nullable": true
                    },
                    "workbench_url": {
                        "type": "string",
                        "readOnly": true
                    },
                    "match_candidates": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "additionalProperties": {}
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "available_edition_id",
                    "available_work_id",
                    "cataloging_session_id",
                    "cover_url",
                    "file_status",
                    "kind",
                    "match_candidates",
                    "reader_url",
                    "status",
                    "work_url",
                    "workbench_url"
                ]
            },
            "IssueItemKindEnum": {
                "enum": [
                    "catalog",
                    "planned"
                ],
                "type": "string",
                "description": "* `catalog` - catalog\n* `planned` - planned"
            },
            "IssueItemRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "kind": {
                        "$ref": "#/components/schemas/IssueItemKindEnum"
                    },
                    "work_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "edition_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "title": {
                        "type": "string"
                    },
                    "authors": {
                        "type": "string"
                    },
                    "version_note": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "doi": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "position": {
                        "type": "integer"
                    },
                    "document_type": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "kind"
                ]
            },
...
            "RecommendationIssue": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "issue_label": {
                        "type": "string"
                    },
                    "introduction": {
                        "type": "string"
                    },
                    "public_byline": {
                        "type": "string"
                    },
                    "body_blocks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IssueBodyBlock"
                        }
                    },
                    "cover_url": {
                        "type": "string"
                    },
                    "cover_rendition_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "display_from": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IssueItem"
                        }
                    },
                    "edit_version": {
                        "type": "string"
                    },
                    "draft_revision_id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "nullable": true
                    },
                    "has_unpublished_changes": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "draft_revision_id",
                    "has_unpublished_changes",
                    "id",
                    "published_at",
                    "title"
                ]
            },
            "RecommendationIssueCollection": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true
                    },
                    "current": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RecommendationIssue"
                            }
                        ],
                        "nullable": true
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecommendationIssue"
                        }
                    }
                },
                "required": [
                    "count",
                    "current",
                    "next",
                    "previous",
                    "results"
                ]
            },
            "RecommendationIssueRequest": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string",
                        "minLength": 1
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "issue_label": {
                        "type": "string"
                    },
                    "introduction": {
                        "type": "string"
                    },
                    "public_byline": {
                        "type": "string"
                    },
                    "body_blocks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IssueBodyBlockRequest"
                        }
                    },
                    "cover_url": {
                        "type": "string"
                    },
                    "cover_rendition_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "display_from": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IssueItemRequest"
                        }
                    },
                    "edit_version": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "title"
                ]
            },
...
            "TypeEnum": {
                "enum": [
                    "paragraph",
                    "heading",
                    "quote",
                    "link"
                ],
                "type": "string",
                "description": "* `paragraph` - paragraph\n* `heading` - heading\n* `quote` - quote\n* `link` - link"
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "libraryCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "stl_access"
            }
        }
    }
}